Split XmlResultParser
